% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots_g.R
\name{boots_g}
\alias{boots_g}
\title{Bootstrap standard errors}
\usage{
boots_g(seqg, boots = 1000)
}
\arguments{
\item{seqg}{Output from sequential_g}

\item{boots}{The number of bootstrap replicates. Defaults to 1000.}
}
\description{
Called internally in sequential_g
}
\examples{

data(ploughs)
form <- women_politics ~ plow +
 agricultural_suitability + tropical_climate + large_animals + rugged | 
 years_civil_conflict + years_interstate_conflict  + oil_pc +
 european_descent + communist_dummy + polity2_2000 | 
 centered_ln_inc + centered_ln_incsq
s1 <- sequential_g(form, ploughs)
 
out.boots <- boots_g(s1)

out.boots

}
