% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw.R
\name{cde_aipw}
\alias{cde_aipw}
\title{Initialize an AIPW CDE estimator}
\usage{
cde_aipw(trim = c(0.01, 0.99), aipw_blip = TRUE)
}
\arguments{
\item{trim}{A vector of length 2 indicating what quantiles of the
propensity scores should be trimmed. By default this is \code{c(0.01, 0.99)} meaning that the top and bottom 1\% of propensity scores are
trunctated to these quantiles. If NULL, no trimming occurs.}

\item{aipw_blip}{If \code{TRUE} (the default), augmented inverse probability weighting
estimators will be used to estimate intermediate outcome
regressions (blip functions).}
}
\description{
Initializes the specification of a CDE estimator based on an
augmented inverse probability weighting approach.
}
