% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw.R
\name{cde_ipw}
\alias{cde_ipw}
\title{Initialize an IPW CDE estimator}
\usage{
cde_ipw(hajek = TRUE, trim = c(0.01, 0.99))
}
\arguments{
\item{hajek}{If \code{TRUE}, normalized weights will be used as in the
Hajek estimator. If \code{FALSE}, traditional IPW weights will be used.}

\item{trim}{A vector of length 2 indicating what quantiles of the
propensity scores should be trimmed. By default this is \code{c(0.01, 0.99)} meaning that the top and bottom 1\% of propensity scores are
trunctated to these quantiles. If NULL, no trimming occurs.}
}
\description{
Initializes the specification of a CDE estimator based on an
inverse probability weighting approach.
}
