% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_impute.R
\name{cde_reg_impute}
\alias{cde_reg_impute}
\title{Initialize an regression imputation CDE estimator}
\usage{
cde_reg_impute(...)
}
\arguments{
\item{...}{Optional arguments to pass to the regression imputation
estimator.}
}
\description{
Initializes the specification of a CDE estimator based on an
regression imputation approach
}
