% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telescope_match.R
\name{cde_telescope_match}
\alias{cde_telescope_match}
\title{Initialize an telescope matching CDE estimator}
\usage{
cde_telescope_match(...)
}
\arguments{
\item{...}{Optional arguments to pass to the telescope matching
estimator.}
}
\description{
Initializes the specification of a CDE estimator based on an
telescope matching approach
}
