% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectEffects.R
\name{seq.g.vcov}
\alias{seq.g.vcov}
\title{Consistent Variance estimator}
\usage{
\method{seq}{g.vcov}(first_mod, direct_mod, X1, X2, med.vars)
}
\arguments{
\item{first_mod}{first stage model}

\item{direct_mod}{DirectEffects model with terms}

\item{X1}{model matrix of first stage model}

\item{X2}{model matrix of DirectEffects model}

\item{med.vars}{vector of term.label attributes for mediators}
}
\description{
Called internally in sequential_g
}
\keyword{internal}
