% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.seqg.R
\name{summary.seqg}
\alias{summary.seqg}
\title{Computes standard errors and p-values of DirectEffects estimates}
\usage{
\method{summary}{seqg}(object, ...)
}
\arguments{
\item{object}{An object of class \env{seqg}, computed by
\code{\link{sequential_g}}.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Computes standard errors and p-values of DirectEffects estimates
}
