% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.seqg.R
\name{summary.seqgboots}
\alias{summary.seqgboots}
\title{Summary of DirectEffect Bootstrap Estimates}
\usage{
\method{summary}{seqgboots}(object, level = 0.95, ...)
}
\arguments{
\item{object}{An output of class \code{seqg} estimated by \code{\link{boots_g}}.}

\item{level}{level of intervals to estimate. Defaults to 0.95}
}
\description{
Summary of DirectEffect Bootstrap Estimates
}
