% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telescope_match.R
\name{telescope_match}
\alias{telescope_match}
\title{Perform telescope matching to estimate the controlled
direct effect of a binary treatment net the effect of a binary mediator.}
\usage{
telescope_match(
  outcome,
  treatment,
  mediator,
  s1.formula,
  s2.formula,
  data,
  caliper = NULL,
  L = 5,
  boot = F,
  nBoot = 5000,
  ci = 95,
  verbose = T
)
}
\arguments{
\item{outcome}{character indicating the name of the
variable in \code{data} being used as the outcome. This variable must appear
on the left-hand side of both \code{s1.formula} and \code{s2.formula}.}

\item{treatment}{character indicating the name of the
variable in \code{data} being used as the treatment indicator. This variable
must be a binary integer (either 0 or 1). This variable must appear
on the right-hand side of  \code{s1.formula}. It should also appear in \code{s2.formula}.}

\item{mediator}{character indicating the name of the
variable in \code{data} being used as the mediator indicator. This variable
must be a binary integer (either 0 or 1). This method returns controlled direct effects
of setting the mediator to 0. Recode the indicator if the controlled direct effect of setting the 
mediator to 1 is of interest. This variable cannot appear in either
\code{s1.formula} or \code{s2.formula}.}

\item{s1.formula}{A formula object denoting the stage 1 formula. The outcome should appear on the left-hand side.
Treatment, pre- and post-treatment covariates should appear on the right-hand side. The mediator is omitted as the model
is fit only within the subset of observations with \code{mediator} equal to 0.}

\item{s2.formula}{A formula object denoting the stage 2 formula estimating the ACDE of treatment using the demediated
matches from the first stage. The outcome should appear on the left-hand side.
Treatment and pre-treatment covariates should appear on the right-hand side.}

\item{data}{A dataframe containing columns referenced by
\code{outcome}, \code{treatment} and \code{mediator} along with any variables
referenced in \code{s1.formula} and \code{s2.formula}.}

\item{caliper}{A scalar denoting the caliper to be used in matching in the treatment stage (calipers cannot be used for matching
on the mediator). Observations outside of the caliper are dropped. Calipers are specified in standard deviations of the covariates.
NULL by default (no caliper).}

\item{L}{Number of matches to use for each unit. Must be a numeric vector of eitehr length 1 or 2. If length 1, L 
sets the number of matches used in both the first stage (matching on mediator) and in the second stage (matching on treatment).
If length 2, the first element sets the number of matches used in the first stage (matching on mediator) and the second element
sets the number of matches used in the second stage (matching on treatment) Default is 5.}

\item{boot}{logical indicating whether to conduct inference using the weighted bootstrap 
for matching estimators extended from Otsu and Rai (2017) (\code{TRUE}) or the asymptotic variance 
estimator from Blackwell and Strezhnev (2019) (\code{FALSE}). Defaults to \code{FALSE}.}

\item{nBoot}{If \code{boot} is \code{TRUE}, number of bootstrap iterations to use. Default is \code{5000}.}

\item{ci}{percent level of confidence interval to return. If \code{boot} is \code{FALSE}, returns symmetric asymptotic 
interval centered on the estimated treatment effect. If \code{boot} is \code{TRUE} returns the \code{(100 - ci)/2} and
\code{100 - (100 - ci)/2} percentiles of the bootstrap distribution. Must be in the interval \code{(0, 100)}. Defaults to 95.}

\item{verbose}{logical indicating whether to display progress information. Default is \code{TRUE}.}
}
\value{
Returns an object of \code{class} \code{tmatch}. Contains the following components
\itemize{
\item estimate: Estimated ACDE fixing M=0
\item std.err: Estimated asymptotic standard error. \code{NULL} if \code{boot} is \code{TRUE}
\item boot.dist: Bootstrap distribution of \code{estimate}. \code{NULL} if \code{boot} is \code{FALSE}
\item conf.low: Lower bound of \code{ci} confidence interval for the estimate
\item conf.high: Upper bound of \code{ci} confidence interval for the estimate
\item ci.level: Level of the confidence interval
\item outcome: Name of outcome variable
\item treatment: Name of treatment variable
\item mediator: Name of mediator variable
\item pre.treatment: Vector of names of pre-treatment confounders (appear in both stage 1 and 2)
\item post.treatment: Vector of names of post-treatment confounders (appear only in stage 1)
\item s1.formula: Stage 1 bias-correction regression formula (pre-/post-treatment covariates)
\item s2.formula: Stage 2 bias-correction regression formula (pre-treatment covariates)
\item outcome.vec: Vector of outcomes used in estimation
\item treatment.vec: Vector of treatment indicators used in estimation
\item mediator.vec: Vector of mediator indicators used in estimation
\item L_m: Number of matches found for each unit in the first stage mediator matching procedure
\item L_a: Number of matches found for each unit in the second stage mediator matching procedure
\item KLm: Number of times unit is used as a match in the first stage mediator matching procedure
\item KLa: Number of times unit is used as a match in the second stage treatment matching procedure
\item N: Number of observations
\item N_summary: Number of observations in each treatment/mediator combination.
\item caliper: Caliper (if any) used in the treatment stage matching to drop distant observations.
}
}
\description{
Perform telescope matching to estimate the controlled
direct effect of a binary treatment net the effect of a binary mediator.
}
\details{
The \code{telescope_match} function implements the two-stage
"telescope matching" procedure developed by Blackwell and Strezhnev 
(2019).

 The procedure first estimates a demediated outcome using a combination
 of matching and a regression bias-correction. The first stage formula 
 \code{s1.formula} specifies the pre- and post-treatment covariates to be used
 in matching along with the specification for the bias-correction regression. In this stage,
 all units with M = 1 to units with M = 0 with identical treatment values and 
 similar pre- and post-treatment covariates. The potential outcomes under M = 0
 are imputed using the average of the matches + the bias correction from the regression
 model. The second stage estimates the controlled direct effect of treatment
 on this demediated outcome using a similar matching/bias-correction procedure with
 the formula specified in \code{s2.formula} indicating the pre-treatment covariates used
 along with the treatment.
 
 Matching is performed using the \code{Match()} routine from the \code{Matching} package.
 By default, matching is L-to-1 nearest neighbor with replacement using Mahalanobis distance.
 

See the references below for more details.
}
\examples{
data(jobcorps)

## Split male/female
jobcorps_female <- jobcorps \%>\% filter(female == 1)

## Telescope matching formula - First stage (X and Z)
tm_stage1 <- exhealth30 ~ treat*(schobef + trainyrbef + jobeverbef + jobyrbef + health012 + health0mis +  pe_prb0 + 
                                   everalc + alc12 + everilldrugs + age_cat +  eduhigh + rwhite + everarr + hhsize + hhsizemis +  hhinc12 + hhinc8 + fdstamp +
                                   welf1 + welf2 + publicass + emplq4 + emplq4full + pemplq4 + pemplq4mis + vocq4 + vocq4mis + 
                                   health1212 + health123 + pe_prb12 + pe_prb12mis  + 
                                   narry1 + numkidhhf1zero + numkidhhf1onetwo + pubhse12 + h_ins12a + h_ins12amis)

## Telescope matching formula - second stage (X)
tm_stage2 <- exhealth30 ~ treat*(schobef + trainyrbef + jobeverbef + jobyrbef + health012 + health0mis +  pe_prb0 + 
                                   everalc + alc12 + everilldrugs + age_cat +  eduhigh +  rwhite + everarr + hhsize + hhsizemis + hhinc12 + hhinc8 + fdstamp +
                                   welf1 + welf2 + publicass)


### Estimate ACDE for women holding employment at 0
telescopeMatch.result.0 <-  telescope_match(outcome = "exhealth30", treatment = "treat", mediator = "work2year2q", 
                                           s1.formula = tm_stage1, 
                                           s2.formula = tm_stage2, data=jobcorps_female, L=3, boot=F, nBoot=1000, verbose=T, ci=95)

}
\references{
Blackwell, Matthew, and Strezhnev, Anton (2020)
"Telescope Matching: Reducing Model Dependence 
in the Estimation of Direct Effects." Working Paper.
}
