% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telescope_match.R
\name{tmatch_weights}
\alias{tmatch_weights}
\title{Calculate telescope matching weights}
\usage{
tmatch_weights(object, data)
}
\arguments{
\item{object}{an object of class `tmatch` from [telescope_match]
output.}

\item{data}{a data.frame that was passed to [telescope_match] to
produce `object`.}
}
\value{
A list containing two vectors:
\item{treat_w}{A n-length vector containing the matching weights
for the treatment or first period matching.}
\item{med_w}{A n-length vector containing the matching weights
for the mediator or second period matching.}
}
\description{
Calculates the weight given to each unit in the telescope matching
estimation procedure.
}
\author{
Matthew Blackwell
}
