% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots_g.R
\name{boots_g}
\alias{boots_g}
\title{Bootstrap standard errors}
\usage{
boots_g(seqg, boots = 1000)
}
\arguments{
\item{seqg}{Output from sequential_g}

\item{boots}{The number of bootstrap replicates. Defaults to 1000.}
}
\description{
Called internally in sequential_g
}
