% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectEffects.R, R/boots_g.R
\name{boots_g}
\alias{boots_g}
\alias{boots_g}
\title{Bootstrap standard errors}
\usage{
boots_g(seqg, boots = 1000, progress = TRUE)

boots_g(seqg, boots = 1000, progress = TRUE)
}
\arguments{
\item{seqg}{Output from sequential_g}

\item{boots}{The number of bootstrap replicates. Defaults to 1000.}

\item{progress}{Show a textbar for progress?}

\item{seqg}{Output from sequential_g}

\item{boots}{The number of bootstrap replicates. Defaults to 1000.}

\item{progress}{Show a textbar for progress?}
}
\value{
\item{acde.sd}{a vector of bootstrapped standard errors for each named coefficient for the direct effects model.}
\item{ate.sd}{a vector of bootstrapped stanadard errors for each named coefficient for the
average treatment effects effects model.}
\item{pt.sd}{a vector of bootstrapped stanadard errors for each named coefficient for the
first stage effects model.}
}
\description{
Called internally in sequential_g

Estimate bootstrap standard errors
}
\examples{
data(civilwar)

# First stage
fit_first <- lm(onset ~ ethfrac + warl + gdpenl + lpop +
                  ncontig + Oil + nwstate + polity2l + relfrac + instab,
                data = civilwar)


# main formula: Y ~ A + X | M
form_main <- onset ~ ethfrac + lmtnest + ncontig + Oil | instab

# estimate CDE
direct <- sequential_g(form_main, fit_first, data = civilwar)

# bootstrap
out_boot <- boots_g(direct)

# print bootstrapped standard errors
print(out_boot)

}
\keyword{internal}
