% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots_g.R
\name{boots_g}
\alias{boots_g}
\title{Bootstrap Standard Errors for Sequential G model object}
\usage{
boots_g(seqg, boots = 1000, showbar = TRUE)
}
\arguments{
\item{seqg}{Output from sequential_g}

\item{boots}{The number of bootstrap replicates. Defaults to 1000.}

\item{showbar}{Show a textbar for progress?}
}
\value{
\item{acde.sd}{a vector of bootstrapped standard errors for each named coefficient for the direct effects model.}
\item{ate.sd}{a vector of bootstrapped stanadard errors for each named coefficient for the
average treatment effects effects model.}
\item{pt.sd}{a vector of bootstrapped stanadard errors for each named coefficient for the
first stage effects model.}
}
\description{
Estimate bootstrap standard errors
}
\examples{
data(civilwar)

# First stage
fit_first <- lm(onset ~ ethfrac + warl + gdpenl + lpop +
                  ncontig + Oil + nwstate + polity2l + relfrac + instab,
                data = civilwar)

rows_use <- rownames(civilwar) \%in\% rownames(model.matrix(fit_first)) # listwise deletion

# main formula: Y ~ A + X | M
form_main <- onset ~ ethfrac + lmtnest + ncontig + Oil | instab

# estimate CDE
direct <- sequential_g(form_main, fit_first, data = civilwar, subset = rows_use)

# bootstrap
out_boot <- boots_g(direct)

# print bootstrapped standard errors
print(out_boot)

}
