% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{cdesens}
\alias{cdesens}
\title{Sensitivity of ACDE}
\usage{
cdesens(seqg, trvar, medvar, rho = seq(-0.9, 0.9, by = 0.05))
}
\arguments{
\item{seqg}{Output from sequential_g}

\item{trvar}{character vector for treatment (abstract away later)}

\item{medvar}{character vector for mediator (abstract away later, allowing for more than 1)}

\item{rho}{A numerical vector of correlations between errors to test for. The 
original model asusmes \env{rho = 0}}
}
\description{
Estimate how ACDE varies by various levels of unmeasured confounding
}
\examples{
data(fear)  
# First stage
fit_first <- lm(onset ~ ethfrac + warl + gdpenl + lpop + 
                  ncontig + Oil + nwstate + polity2l + relfrac + instab,
                data = fear) # lm(Y ~ A + X + M)

rows_use <- rownames(fear) \%in\% rownames(model.matrix(fit_first)) # listwise deletion

# main formula
form_main <- onset ~ ethfrac + lmtnest + ncontig + Oil | instab # Y ~ A + X | M

# estimate CDE
direct <- sequential_g(form_main, fit_first, data = fear, subset = rows_use)

# sensitivity 
out_sens <- cdesens(direct, trvar = "ethfrac", medvar = "instab")

# show sensitivity
plot(out_sens)

}
