% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.R
\name{plot.cdesens}
\alias{plot.cdesens}
\title{Plot output from cdesens}
\usage{
\method{plot}{cdesens}(x, level = 0.975, ...)
}
\arguments{
\item{x}{output from \env{cdesens}}
}
\description{
Plot output from cdesens
}
