% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telescope_match.R
\name{plotDiag.tmatch}
\alias{plotDiag.tmatch}
\title{Histograms of matching weights}
\usage{
plotDiag.tmatch(object, stage = "mediator")
}
\arguments{
\item{object}{an object of class \code{tmatch} -- results from a
call  to \code{telescope_match}}

\item{stage}{a character vector equal to either 'mediator' or
'treatment'. If equal to 'mediator', returns a histogram
of matching weights for units with mediator = 0. If equal to
'treatment', returns a histogram of matching weights for
all units.}
}
\value{
Outputs a `plot()` object containing the histogram of match counts
}
\description{
Histograms of matching weights
}
\details{
Provides histograms of the number of times each unit is
used as a match given a  \code{tmatch} object returned by
\code{telescope_match}
}
