% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telescope_match.R
\name{summary.tmatch}
\alias{summary.tmatch}
\title{Summarize telescope match objects}
\usage{
\method{summary}{tmatch}(object)
}
\arguments{
\item{object}{an object of class \code{tmatch} -- results from a call to \code{telescope_match}}
}
\description{
Summarize telescope match objects
}
\details{
\code{summary} method for \code{tmatch} objects returned by
\code{telescope_match}
}
