% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telescope_match.R
\name{telescope_match}
\alias{telescope_match}
\title{Perform telescope matching to estimate the controlled
direct effect of a binary treatment net the effect of a binary mediator.}
\usage{
telescope_match(
  formula,
  data,
  caliper = NULL,
  L = 5,
  boot = FALSE,
  nBoot = 5000,
  ci = 95,
  verbose = TRUE,
  subset,
  contrasts = NULL,
  separate_bc = TRUE,
  ...
)
}
\arguments{
\item{formula}{A formula object that specifies the outcome,
baseline covariates, treatment, intermediate covariate, and
mediator to be used in the matching. Each of the last four
variable groups should be separated by \code{|}. See below for
more details.}

\item{data}{A dataframe containing columns referenced by
\code{formula}.}

\item{caliper}{A scalar denoting the caliper to be used in matching
in the treatment stage (calipers cannot be used for matching on
the mediator). Observations outside of the caliper are dropped.
Calipers are specified in standard deviations of the covariates.
NULL by default (no caliper).}

\item{L}{Number of matches to use for each unit. Must be a numeric
vector of eitehr length 1 or 2. If length 1, L sets the number of
matches used in both the first stage (matching on mediator) and
in the second stage (matching on treatment). If length 2, the
first element sets the number of matches used in the first stage
(matching on mediator) and the second element sets the number of
matches used in the second stage (matching on treatment) Default
is 5.}

\item{boot}{logical indicating whether to conduct inference using
the weighted bootstrap for matching estimators extended from Otsu
and Rai (2017) (\code{TRUE}) or the asymptotic variance estimator
from Blackwell and Strezhnev (2019) (\code{FALSE}). Defaults to
\code{FALSE}.}

\item{nBoot}{If \code{boot} is \code{TRUE}, number of bootstrap
iterations to use. Default is \code{5000}.}

\item{ci}{percent level of confidence interval to return. If
\code{boot} is \code{FALSE}, returns symmetric asymptotic
interval centered on the estimated treatment effect. If
\code{boot} is \code{TRUE} returns the \code{(100 - ci)/2} and
\code{100 - (100 - ci)/2} percentiles of the bootstrap
distribution. Must be in the interval \code{(0, 100)}. Defaults
to 95.}

\item{verbose}{logical indicating whether to display progress
information. Default is \code{TRUE}.}

\item{subset}{A vector of logicals indicating which rows of
\code{data} to keep.}

\item{contrasts}{a list to be passed to the \code{contrasts.arg}
argument of \code{model.matrix()} when generating the data matrix.}

\item{separate_bc}{logical indicating whether or not bias
correction regressions should be run separately within levels of
the treatment and mediator. Defaults to \code{TRUE}. If
\code{TRUE}, any interactions between treatment/mediator and
covariates in the specification should be omitted.}

\item{...}{additional arguments to be passed to the low level
    regression fitting functions (see below).}
}
\value{
Returns an object of \code{class} \code{tmatch}. Contains
the following components
\itemize{
\item estimate: Estimated ACDE fixing M=0
\item std.err: Estimated asymptotic standard error. \code{NULL} if
\code{boot} is \code{TRUE}
\item boot.dist: Bootstrap distribution of \code{estimate}.
\code{NULL} if  \code{boot} is \code{FALSE}
\item conf.low: Lower bound of \code{ci} confidence interval for
the estimate
\item conf.high: Upper bound of \code{ci} confidence interval for
the estimate
\item ci.level: Level of the confidence interval
\item outcome: Name of outcome variable
\item treatment: Name of treatment variable
\item mediator: Name of mediator variable
\item pre.treatment: Vector of names of pre-treatment confounders
(appear in both stage 1  and 2)
\item post.treatment: Vector of names of post-treatment confounders
(appear only in stage 1)
\item s1.formula: Stage 1 bias-correction regression formula
(pre-/post-treatment covariates)
\item s2.formula: Stage 2 bias-correction regression formula
(pre-treatment covariates)
\item outcome.vec: Vector of outcomes used in estimation
\item treatment.vec: Vector of treatment indicators used in
estimation
\item mediator.vec: Vector of mediator indicators used in
estimation
\item L_m: Number of matches found for each unit in the first stage
mediator matching  procedure
\item L_a: Number of matches found for each unit in the second
stage mediator matching  procedure
\item KLm: Number of times unit is used as a match in the first
stage mediator matching  procedure
\item KLa: Number of times unit is used as a match in the second
stage treatment matching  procedure
\item N: Number of observations
\item N_summary: Number of observations in each treatment/mediator
combination.
\item caliper: Caliper (if any) used in the treatment stage
matching to drop distant  observations.
}
}
\description{
Perform telescope matching to estimate the controlled
direct effect of a binary treatment net the effect of a binary mediator.
}
\details{
The \code{telescope_match} function implements the
  two-stage "telescope matching" procedure developed by Blackwell
  and Strezhnev (2020).

 The procedure first estimates a demediated outcome using a
 combination of matching and a regression bias-correction based on
 the specification in \code{formula}, which should be specified as
 \code{Y ~ X | A | Z | M}, where \code{Y} is the outcome, \code{X}
 is a formula of baseline covariates, \code{A} is a single variable
 name indicating the binary treatment, \code{Z} is a formula of
 intermediate covariates, and \code{M} is a single variable name
 indicating the mediator.

Under the default \code{separate_bc == TRUE}, the first stage will
match for \code{M} on \code{X} and \code{Z} within levels of
\code{A}. The bias correction regressions will be linear with a
specification of \code{Y ~ X + Z} within levels of \code{A} and
\code{M}. In the second stage, we match for \code{A} within levels
of \code{X} and run bias correction regressions of \code{Ytilde ~
X} within levels of \code{A}, where \code{Ytilde} is the imputed
value of \code{Y(A, 0)} from the first stage. This second step
estimates the controlled direct effect of treatment when the
mediator is 0.

When \code{separate_bc} is \code{FALSE}, the bias correction
regressions are not broken out by the treatment/mediator and
\code{A} and \code{M} are simply included as separate terms in the
linear regression. In this setting, interactions between the
treatment/mediator and covariates can be added on a selective basis
to the \code{X} and \code{Z} specifications.

 Matching is performed using the \code{Match()} routine from the
\code{Matching} package. By default, matching is L-to-1 nearest
neighbor with replacement using Mahalanobis distance.


See the references below for more details.
}
\examples{
data(jobcorps)

## Split male/female
jobcorps_female <- subset(jobcorps, female == 1)

## Telescope matching formula - First stage (X and Z)
tm_form <- exhealth30 ~  schobef + trainyrbef + jobeverbef  |
treat | emplq4 + emplq4full | work2year2q


### Estimate ACDE for women holding employment at 0
tm_out <-  telescope_match(
  tm_form,
  data = jobcorps_female,
  L = 3,
  boot = FALSE,
  verbose = TRUE
)

}
\references{
Blackwell, Matthew, and Strezhnev, Anton (2020)
"Telescope Matching: Reducing Model Dependence
in the Estimation of Direct Effects." Working Paper.
}
